Makesense {geneplotter}	R Documentation

Produce Smoothed Sense/Anti-sense For All Chromosomes

Description

'Makesense' takes either an ExpressionSet object or a matrix of gene expressions and will produce a smoothed positive and negative strands for all chromosomes.

Usage


Makesense(expr, lib, ...)

Arguments

expr
Either an ExpressionSet or a matrix of gene expressions with genes as rows and columns as samples.

lib
The name of the Bioconductor annotation data package that will be used to provide mappings from probes to chromosomal locations, such as hgu95av2.db or hgu133a.db. If expr is an ExpressionSet, the argument defaults to the annotation slot of the ExpressionSet.

...
Currently, the only optional argument is f, the smoother span to be passed to 'lowess'. Its value should be in the interval of (0,1). This gives the proportion of points in the plot which influence the smooth at each value. Larger values give more smoothness. The default value for this argument is 1/10.

Details

The expr argument can either be of class ExpressionSet or matrix, where the latter represents the matrix of gene expressions.

If the expr argument is an ExpressionSet, the lib argument will use the annotation slot. Users can override this behaviour and supply their own lib argument if they wish. If the ExpressionSet has no value associated with the annotation slot (which should not happen, but is possible) then the user must supply the lib argument manually or the function will throw an error.

Value

A list of 2 components:

ans2
a list, whose components correspond to samples in the same order as appearing in the columns of 'expr'. Each component is also a list, named by chromosomes "1"-"22", "X" and "Y". Each named component is again a list with two elements named "posS" and "negS", corresponding to the positive and negative strands of a chromosome, each of which is an object returned by 'lowess'.

lib
A string giving the name of the annotation data package to use. Optional if expr is an ExpressionSet.

Author(s)

Robert Gentleman and Xiaochun Li

See Also

plotChr

Examples


  if (require("hgu133a.db")) {
    data(expressionSet133a)
    esetobj <- Makesense(exprs(expressionSet133a), "hgu133a")
    esetobj2 <- Makesense(expressionSet133a[1:200, ])
  }

[Package geneplotter version 1.60.0 Index]