amplicon.plot {geneplotter}	R Documentation

Create an amplicon plot

Description

Given a two-sample test statistic and an ExpressionSet this function plots regions of the genome that are either highly expressed (in red) or have low expression (blue) differentially in the two groups.

Usage


amplicon.plot(ESET, FUN, genome)

Arguments

ESET
an object of class ExpressionSet

FUN
A two sample test function suitable for esApply.

genome
A character string of the base name for the annotation.

Details

In some genetic studies we are interested in finding regions of the genome where there are a set of highly expressed genes in some subgroup of the population. This set of highly (or lowly) expressed genes is often of great interest. For example in breast cancer the HER–2 gene is on an amplicon. In some patients approximately 5 genes located near HER–2 are all amplified.

These plot should help in the search for such regions.

Value

No value is returned. This function is executed purely for side effect.

Author(s)

Robert Gentleman

See Also

esApply, make.chromOrd

Examples


  ##none yet; takes too long

[Package geneplotter version 1.60.0 Index]