cColor {geneplotter}	R Documentation

A function for marking specific probes on a cPlot.

Description

Given a set of probes, will highlight them in the color desired on a plot which has already been created via the function cPlot().

Usage


cColor(probes, color, plotChroms, scale=c("relative","max"), glen=0.4,
       ...)

Arguments

probes
The probes that are being highlighted.

color
A vector of colors, recycled as necessary, to highlight the probes.

plotChroms
An object of type chromLocation which contains all the gene information to be plotted.

scale
Whether to plot the graph scaled absolutely or relative by chromosome. Default is absolute.

glen
The length of the gene line plotted.

...
Additional graphics arguments, passed to segments, which is used to draw the vertical ticks.

Details

It is important to call the function cPlot() first. This function will then search for the specific locations of the probes desired, which are contained within the plotChroms instance of a chromLocation class. It will then pass these on to the plotting routine to highlight the desired locations. NOTE: It is important that plotChroms, scale and glen parameters are the same as used for cPlot().

Author(s)

Jeff Gentry

See Also

cPlot, chromLocation-class

Examples


  if (require("hgu95av2.db")) {
    z <- buildChromLocation("hgu95av2")
    cPlot(z)
    probes <- c("266_s_at", "31411_at", "610_at", "failExample")
    cColor(probes, "red", z)
    probes2 <- c("960_g_at", "41807_at", "931_at", "39032_at")
    cColor(probes2, "blue", z)
  } else
    print("Need hgu95av2.db data package for the example")
  

[Package geneplotter version 1.60.0 Index]