cScale {geneplotter}	R Documentation

A function for mapping chromosome length to a number of points.

Description

Given a number of points (generally representing the number of points on a plot's axis), and a vector of chromosome lengths - will generate a vector of the same length as the one passed in containing scaling factors for each chromosome.

Usage


cScale(points, cLengths, method=c("max", "relative"), chrom)

Arguments

points
The number of points to scale the chromosome length to.

cLengths
A vector of chromosome lengths.

method
Determines whether to use relative or absolute scaling. Default is "max" (absolute).

chrom
Which chrom to determine the scale for

Details

The scale factor is calculated in a manner based on the method argument. If method is max, the factor is derived by dividing the points argument by each chromosome's length (in base pairs). If the method chosen is relative, then the scale is determined by dividing the points argument by the maximum chromsome length, and applying that value to each chromosome.

Author(s)

Jeff Gentry

See Also

cPlot

Examples


   ## A bit of a hack to not have a package dependency on hgu95av2
   ## but need to fiddle w/ the warn level to not fail the example anyways.
   curWarn <- options(warn=0)
   on.exit(options(warn), add=TRUE)
   if (require("hgu95av2.db")) {
      z <- buildChromLocation("hgu95av2")

     for (sc in c("max","relative"))
       scale <- cScale(1000, chromLengths(z),sc,"Y")
   } else print("This example needs the hgu95av2 data package")

[Package geneplotter version 1.60.0 Index]