groupedHeatmap {geneplotter}	R Documentation

Heatmap of a matrix with grouped rows and columns

Description

The function uses grid.rect and grid.rect to draw a heatmap with grouped rows and columns.

Usage


groupedHeatmap(z, frow, fcol,
  fillcolours = c("#2166ac","#4393c3","#92c5de","#d1e5f0","#fefefe","#fddbc7","#f4a582","#d6604d","#b2182b"),
  bordercolour = "#e0e0e0",
  zlim = range(z, na.rm=TRUE))

Arguments

z
A matrix with row and column names.

frow
A factor of length nrow(z) indicating the row grouping.

fcol
A factor of length ncol(z) indicating the column grouping.

fillcolours
A character vector of colours from which the colour map is obtained through interpolation.

bordercolour
Either a character vector of length 1, specifying the border colour of the heatmap tiles, or NULL or NA, which indicates that the border colour should match the fill colour.

zlim
Lower and upper limit of z values represented in the colour map.

Details

The function can be called within other drawing operations from the grid package, e.g. within a viewport.

Value

The function is called for its side effect, drawing text and rectangles on the current viewport.

Author(s)

Wolfgang Huber http://www.ebi.ac.uk/huber

See Also

grid.text, grid.rect

Examples



data("mtcars")

groupedHeatmap(
  scale(mtcars),
  frow = factor(sapply(strsplit(rownames(mtcars), " "), "[", 1)),
  fcol = factor(round(seq_len(ncol(mtcars))/3)))

[Package geneplotter version 1.60.0 Index]