make.chromOrd {geneplotter}	R Documentation

Make a chromOrd object

Description

This function makes a chromOrd object.

Usage


make.chromOrd(genome, gnames)

Arguments

genome
A character string.

gnames
A character vector of the genes to be selected.

Details

This function reads in a lot of annotation data and creates a list with one element for each chromosome. The elements of this list are indices indicating the order of the genes that are on that chromosome (and in the annotation data set being used).

Value

A list of chromOrd type. One element for each chromosome. Suitable for reordering other values according to the chromosomal location.

Author(s)

Robert Gentleman

See Also

amplicon.plot

Examples


  data(sample.ExpressionSet)
  make.chromOrd("hgu95A", featureNames(sample.ExpressionSet))

[Package geneplotter version 1.60.0 Index]