plotMA-methods {geneplotter}	R Documentation

Generate an MA plot

Description

Generate a plot of log fold change versus mean expression (MA plot)

Usage


## S4 method for signature 'data.frame'
plotMA( object, ylim = NULL,
  colNonSig = "gray32", colSig = "red3", colLine = "#ff000080",
  log = "x", cex=0.45, xlab="mean expression", ylab="log fold change", ... )

Arguments

object
A data.frame with (at least) three columns, the first containing the mean expression values (for the x-axis), the second the logarithmic fold change (for the-y axis) and the third a logical vector indicating significance (for the colouring of the dots).

ylim
The limits for the y-axis. If missing, an attempt is made to choose a sensible value. Dots exceeding the limits will be displayed as triangles at the limits, pointing outwards.

colNonSig
colour to use for non-significant data points.

colSig
colour to use for significant data points.

colLine
colour to use for the horizontal (y=0) line.

log
which axis/axes should be logarithmic; will be passed to plot.

cex
The cex parameter for plot.

xlab
The x-axis label.

ylab
The y-axis label.

...
Further parameters to be passed through to plot.

Examples


  plotMA(
    data.frame(
      `M` = exp(rexp(1000)),
      `A` = rnorm(1000) -> tmp,
      `isde` = abs(tmp)>2) 
  )

[Package geneplotter version 1.60.0 Index]